requirements.txt
apache-airflow==2.8.1
pandas==2.0.3
sqlalchemy>=1.4.24,<2.0.0
psycopg2-binary==2.9.9
pydantic>=2.0.3
boto3==1.34.34
pyarrow==10.0.1
fastparquet==2023.10.1
pytest==7.4.3
python-dotenv==1.0.1
great-expectations==0.17.15